add url encode and deocde, separate only with &
